+++ src/main/java/uk/gov/ons/ssdc/supporttool/endpoint/DeactivateUacEndpoint.java
@@ -14,11 +14,11 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
-import uk.gov.ons.ssdc.supporttool.model.dto.DeactivateUacDTO;
-import uk.gov.ons.ssdc.supporttool.model.dto.EventDTO;
-import uk.gov.ons.ssdc.supporttool.model.dto.EventTypeDTO;
-import uk.gov.ons.ssdc.supporttool.model.dto.PayloadDTO;
-import uk.gov.ons.ssdc.supporttool.model.dto.ResponseManagementEvent;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.DeactivateUacDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.EventDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.EventTypeDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.PayloadDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.ResponseManagementEvent;
 import uk.gov.ons.ssdc.supporttool.model.entity.UacQidLink;
 import uk.gov.ons.ssdc.supporttool.model.repository.UacQidLinkRepository;
 import uk.gov.ons.ssdc.supporttool.security.UserIdentity;
