--- src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/EventTypeDTO.java
@@ -1,5 +1,0 @@
-package uk.gov.ons.ssdc.supporttool.model.dto;
-
-public enum EventTypeDTO {
-  DEACTIVATE_UAC
-}
