--- src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/PayloadDTO.java
@@ -1,11 +1,0 @@
-package uk.gov.ons.ssdc.supporttool.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import lombok.Data;
-
-@Data
-@JsonInclude(Include.NON_NULL)
-public class PayloadDTO {
-  private DeactivateUacDTO deactivateUac;
-}
