+++ src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/messaging/DeactivateUacDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ssdc.supporttool.model.dto;
+package uk.gov.ons.ssdc.supporttool.model.dto.messaging;
 
 import lombok.Data;
