+++ src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/messaging/EventTypeDTO.java
@@ -1,0 +1,8 @@
+package uk.gov.ons.ssdc.supporttool.model.dto.messaging;
+
+public enum EventTypeDTO {
+  DEACTIVATE_UAC,
+  ADDRESS_NOT_VALID,
+  REFUSAL_RECEIVED,
+  FULFILMENT
+}
