+++ src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/messaging/PayloadDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ssdc.supporttool.model.messaging.dto;
+package uk.gov.ons.ssdc.supporttool.model.dto.messaging;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -7,6 +7,7 @@ import lombok.Data;
 @Data
 @JsonInclude(Include.NON_NULL)
 public class PayloadDTO {
+  private DeactivateUacDTO deactivateUac;
   private RefusalDTO refusal;
   private InvalidAddressDTO invalidAddress;
   private FulfilmentDTO fulfilment;
