+++ src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/messaging/RefusalDTO.java
@@ -1,16 +1,16 @@
-package uk.gov.ons.ssdc.supporttool.model.ui.dto;
+package uk.gov.ons.ssdc.supporttool.model.dto.messaging;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.RefusalTypeDTO;
 
 @Data
 @NoArgsConstructor
 @JsonInclude(Include.NON_NULL)
-public class Refusal {
+public class RefusalDTO {
   private RefusalTypeDTO type;
   private String agentId;
   private String callId;
+  private CollectionCase collectionCase;
 }
