+++ src/main/java/uk/gov/ons/ssdc/supporttool/model/dto/ui/InvalidAddress.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ssdc.supporttool.model.ui.dto;
+package uk.gov.ons.ssdc.supporttool.model.dto.ui;
 
 import lombok.Data;
