--- src/main/java/uk/gov/ons/ssdc/supporttool/model/messaging/dto/EventTypeDTO.java
@@ -1,16 +1,0 @@
-package uk.gov.ons.ssdc.supporttool.model.messaging.dto;
-
-public enum EventTypeDTO {
-  CASE_CREATED,
-  CASE_UPDATED,
-  UAC_UPDATED,
-  RESPONSE_RECEIVED,
-  REFUSAL_RECEIVED,
-  SAMPLE_LOADED,
-  ADDRESS_NOT_VALID,
-  SURVEY_LAUNCHED,
-  RESPONDENT_AUTHENTICATED,
-  FULFILMENT,
-  TELEPHONE_CAPTURE_REQUESTED,
-  DEACTIVATE_UAC
-}
