+++ src/main/java/uk/gov/ons/ssdc/supporttool/service/CaseService.java
@@ -6,19 +6,19 @@ import java.util.UUID;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.CollectionCase;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.EventDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.EventTypeDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.FulfilmentDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.InvalidAddressDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.PayloadDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.RefusalDTO;
+import uk.gov.ons.ssdc.supporttool.model.dto.messaging.ResponseManagementEvent;
+import uk.gov.ons.ssdc.supporttool.model.dto.ui.Fulfilment;
+import uk.gov.ons.ssdc.supporttool.model.dto.ui.InvalidAddress;
+import uk.gov.ons.ssdc.supporttool.model.dto.ui.Refusal;
 import uk.gov.ons.ssdc.supporttool.model.entity.Case;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.CollectionCase;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.EventDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.EventTypeDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.FulfilmentDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.InvalidAddressDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.PayloadDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.RefusalDTO;
-import uk.gov.ons.ssdc.supporttool.model.messaging.dto.ResponseManagementEvent;
 import uk.gov.ons.ssdc.supporttool.model.repository.CaseRepository;
-import uk.gov.ons.ssdc.supporttool.model.ui.dto.Fulfilment;
-import uk.gov.ons.ssdc.supporttool.model.ui.dto.InvalidAddress;
-import uk.gov.ons.ssdc.supporttool.model.ui.dto.Refusal;
 
 @Service
 public class CaseService {
@@ -69,20 +69,16 @@ public class CaseService {
     RefusalDTO refusalDTO = new RefusalDTO();
     refusalDTO.setCollectionCase(collectionCase);
     refusalDTO.setType(refusal.getType());
-
-    if (refusal.getAgentId() != null) {
-      refusalDTO.setAgentId(refusal.getAgentId());
-    }
-
-    if (refusal.getCallId() != null) {
-      refusalDTO.setCallId(refusal.getCallId());
-    }
+    refusalDTO.setAgentId(refusal.getAgentId());
+    refusalDTO.setCallId(refusal.getCallId());
 
     PayloadDTO payloadDTO = new PayloadDTO();
     payloadDTO.setRefusal(refusalDTO);
 
-    ResponseManagementEvent responseManagementEvent =
-        new ResponseManagementEvent(eventDTO, payloadDTO);
+    ResponseManagementEvent responseManagementEvent = new ResponseManagementEvent();
+
+    responseManagementEvent.setEvent(eventDTO);
+    responseManagementEvent.setPayload(payloadDTO);
 
     rabbitTemplate.convertAndSend(eventsExchange, refusalEventRoutingKey, responseManagementEvent);
   }
@@ -98,16 +94,15 @@ public class CaseService {
     InvalidAddressDTO invalidAddressDTO = new InvalidAddressDTO();
     invalidAddressDTO.setCaseId(caze.getId());
     invalidAddressDTO.setReason(invalidAddress.getReason());
-
-    if (invalidAddress.getNotes() != null) {
-      invalidAddressDTO.setNotes(invalidAddress.getNotes());
-    }
+    invalidAddressDTO.setNotes(invalidAddress.getNotes());
 
     PayloadDTO payloadDTO = new PayloadDTO();
     payloadDTO.setInvalidAddress(invalidAddressDTO);
 
-    ResponseManagementEvent responseManagementEvent =
-        new ResponseManagementEvent(eventDTO, payloadDTO);
+    ResponseManagementEvent responseManagementEvent = new ResponseManagementEvent();
+
+    responseManagementEvent.setEvent(eventDTO);
+    responseManagementEvent.setPayload(payloadDTO);
 
     rabbitTemplate.convertAndSend(
         eventsExchange, invalidAddressEventRoutingKey, responseManagementEvent);
@@ -128,8 +123,10 @@ public class CaseService {
     PayloadDTO payloadDTO = new PayloadDTO();
     payloadDTO.setFulfilment(fulfilmentDTO);
 
-    ResponseManagementEvent responseManagementEvent =
-        new ResponseManagementEvent(eventDTO, payloadDTO);
+    ResponseManagementEvent responseManagementEvent = new ResponseManagementEvent();
+
+    responseManagementEvent.setEvent(eventDTO);
+    responseManagementEvent.setPayload(payloadDTO);
 
     rabbitTemplate.convertAndSend(fulfilmentQueue, responseManagementEvent);
   }
